+++ Code/GCLinearDuplicateController.m
@@ -59,7 +59,7 @@
 }
 
 
-- (void)			sheetDidEnd:(NSWindow *)sheet returnCode:(int)returnCode contextInfo:(void  *)contextInfo
+- (void)sheetDidEnd:(NSWindow *)sheet returnCode:(NSInteger)returnCode contextInfo:(void  *)contextInfo
 {
 #pragma unused (sheet, contextInfo)
 	if ( returnCode == NSOKButton )
